enhance relabel debug
Support for logging metrics subject to relabeling was removed in
v1.85.0 in favor of a web-driven UI. However, this UI is still
not yet practical for inspecting or adjusting scraped metrics.

So if relabel_debug or metric_relabel_debug are set to true in the
sd_configs section, metrics are shown both as scraped and as they
will be sent to the target before finalization, i.e. before
meta tags (labels starting with '__') are removed.

The vmagent flags -remoteWrite.relabelDebug and
-remoteWrite.urlRelabelDebug enable logging of metrics scraped via
-remoteWrite.relabelConfig and -remoteWrite.urlRelabelConfig,
respectively.